First draft.
Actually it was planned to do the first commit with the first
working board, but now the board is done and populated ... and
bugs wre found already. So, make a commit to allow to return here
later.

Board was created mostly with parts from the Gen7 libraries.
V-USB was taken from Gen7's ExtensionBoard USB-A. Some Gen7
heritage may be still left in the sources.