terminal.py: detect and don't print duplicate motor messages.
This used to be handled in the firmware, where this can be done
reliable, but the ATtiny2313 is filled up into the last corner
with more important stuff, so this had to be dropped there.

Doing the same in terminal.py might drop a message or two (if
temperature is exactly the same between two actuations), but
there we have plenty of space for code.